feat: upload支持输出string类型(#52)
